backend/requirements.txt
aiosqlite==0.16.1
autopep8==1.5.5
click==7.1.2
fastapi==0.63.0
h11==0.12.0
iso8601==0.1.14
isort==5.7.0
pycodestyle==2.6.0
pydantic==1.7.3
PyPika==0.44.1
python-dotenv==0.15.0
pytz==2020.5
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
